+++ app/assets/stylesheets/application.css
@@ -15,10 +15,6 @@
  *= require_self
  */
 
-
-
-
-
 body{
 box-sizing: border-box;
 background: #43C6AC;  /* fallback for old browsers */
@@ -32,8 +28,8 @@ background: linear-gradient(to right, #191654, #43C6AC); /* W3C, IE 10+/ Edge, F
   margin-left:40%;
   margin-top: 1%;
   color: black;
-
 }
+
 #form1{
   width: 400px;
 
@@ -41,8 +37,8 @@ background: linear-gradient(to right, #191654, #43C6AC); /* W3C, IE 10+/ Edge, F
   margin-bottom: 2%;
   color: black;
   display:none;
-
 }
+
 .inputfields{
     width: 300px;
     line-height: 10px;
@@ -51,16 +47,20 @@ background: linear-gradient(to right, #191654, #43C6AC); /* W3C, IE 10+/ Edge, F
     margin-bottom: 17px;
     border-style: none;
 }
+
 .margin-bottom{
   margin-bottom: 8px;
 }
+
 .margin-top{
   margin-top: 20px;
 }
+
 .button{
   width: 300px;
    border-radius: 25px;
 }
+
 .description{
   font-size: 10px;
   opacity: 0.7;
