+++ app/controllers/users_controller.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+class UsersController < ApplicationController
+  def index
+    @user = User.all
+  end
+
+  def new
+    @user = User.new
+  end
+
+  def create
+    @user = User.new(user_params)
+    if @user.save
+      redirect_to :index
+    else
+      render :new
+    end
+  end
+
+  def show
+    @user = User.find(params[:id])
+  end
+
+  def user_params
+    params.require(:user).permit(:first_name, :lastname, :email)
+  end
+end
