+++ app/views/devise/registrations/_form.html.erb
@@ -1,5 +1,4 @@
-
-  <div>
+<div>
    <% if resource.errors.any? %>
      <div id="error_explanation">
       <h2 class= "errors" >
@@ -22,38 +21,38 @@
   <%= form_for resource, as: resource_name, url: user_registration_path, method: :post do |f| %>
     <%= render "devise/shared/error_messages", resource: resource %>
 
-    <div class="field">
-      <%= f.label :first_name %><br />
-      <%= f.text_field :firstname, autofocus: true, class: "inputfields", autocomplete: "first_name" %>
-     </div>
+      <div class="field">
+        <%= f.label :first_name %><br />
+        <%= f.text_field :firstname, autofocus: true, class: "inputfields", autocomplete: "first_name" %>
+       </div>
 
-    <div class="field">
-      <%= f.label :last_name %><br />
-      <%= f.text_field :lastname, autofocus: true, class: "inputfields", autocomplete: "last_name" %>
-     </div>
+      <div class="field">
+        <%= f.label :last_name %><br />
+        <%= f.text_field :lastname, autofocus: true, class: "inputfields", autocomplete: "last_name" %>
+       </div>
 
-    <div class="field">
-      <%= f.label :email %><br />
-      <%= f.email_field :email, autofocus: true, class: "inputfields", autocomplete: "email" %>
-    </div>
+      <div class="field">
+        <%= f.label :email %><br />
+        <%= f.email_field :email, autofocus: true, class: "inputfields", autocomplete: "email" %>
+      </div>
 
-    <div class="field">
-      <%= f.label :password %>
-      <% if @minimum_password_length %>
-      <% end %><br />
-      <%= f.password_field :password, class: "inputfields", autocomplete: "new-password" %><br>
-       <em class="description">Password must contain minimum 8 characters, To make it stronger use upper case or lower case letter, numbers and symbols </em>
-    </div>
+      <div class="field">
+        <%= f.label :password %>
+        <% if @minimum_password_length %>
+        <% end %><br />
+        <%= f.password_field :password, class: "inputfields", autocomplete: "new-password" %><br>
+         <em class="description">Password must contain minimum 8 characters, To make it stronger use upper case or lower case letter, numbers and symbols </em>
+      </div>
 
-    <div class="field">
-      <%= f.label :password_confirmation %><br />
-      <%= f.password_field :password_confirmation, class: "inputfields", autocomplete: "new-password" %>
-    </div>
+      <div class="field">
+        <%= f.label :password_confirmation %><br />
+        <%= f.password_field :password_confirmation, class: "inputfields", autocomplete: "new-password" %>
+      </div>
 
-    <div class="actions">
-      <%= f.submit "Sign up" , class: "btn btn-primary margin-bottom button" %>
-    </div>
+      <div class="actions">
+        <%= f.submit "Sign up" , class: "btn btn-primary margin-bottom button" %>
+      </div>
+
   <% end %>
 
   <%= render "devise/shared/links" %>
-
