+++ app/views/devise/registrations/new.js.erb
@@ -1,8 +1,1 @@
-// $(document).ready(function() {
-//   $("#formButton").click(function() {
-//     $("#form1").show();
-//   });
-// });
-
-
-$('.show a').parent().html("<%= j render 'form', resource: resource %>")
+$('.show a').html("<%= j render 'form', resource: resource %>")
