+++ app/views/devise/sessions/_form.html.erb
@@ -1,4 +1,3 @@
-
 <h2>Sign up</h2>
 
 <%= form_for resource, as: resource_name, url: user_registration_path do |f| %>
@@ -35,7 +34,7 @@
   <div class="actions">
     <%= f.submit "Sign up" %>
   </div>
+
 <% end %>
 
 <%= render "devise/shared/links" %>
-
