+++ app/views/devise/sessions/new.html.erb
@@ -1,36 +1,33 @@
-<div class="center">
-  <div class="row1">
-<h1 class="Login">Log in</h1>
+  <div class="center">
+    <div class="row1">
+  <h1 class="Login">Log in</h1>
 
-<%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
-  <div>
-    <%= f.label :email, class: "margin-bottom" %><br />
-    <%= f.email_field :email, autofocus: true, class: "inputfields",  autocomplete: "email" %>
-  </div>
+  <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
+      <div>
+        <%= f.label :email, class: "margin-bottom" %><br />
+        <%= f.email_field :email, autofocus: true, class: "inputfields",  autocomplete: "email" %>
+      </div>
 
-  <div class="inputfields">
-    <%= f.label :password, class: "margin-bottom" %><br />
-    <%= f.password_field :password, class: "inputfields", autocomplete: "current-password" %>
-  </div>
+      <div class="inputfields">
+        <%= f.label :password, class: "margin-bottom" %><br />
+        <%= f.password_field :password, class: "inputfields", autocomplete: "current-password" %>
+      </div>
 
-  <% if devise_mapping.rememberable? %>
-    <div>
-      <%= f.check_box :remember_me %>
-    <label class="form-check-label margin-top" for="exampleCheck1">Remember Me</label>
-    </div>
+      <% if devise_mapping.rememberable? %>
+
+        <div>
+          <%= f.check_box :remember_me %>
+           <label class="form-check-label margin-top" for="exampleCheck1">Remember Me</label>
+        </div>
+
+      <% end %>
+
+      <div class="actions">
+         <%= f.submit "Log in",class: "btn btn-primary margin-bottom button" %>
+      </div>
   <% end %>
 
-  <div class="actions">
-     <%= f.submit "Log in",class: "btn btn-primary margin-bottom button" %>
+  <%= render "devise/shared/links" %>
+
   </div>
-<% end %>
-<%= render "devise/shared/links" %>
-
-
-
- 
-
 </div>
-</div>
-
-
