+++ config/routes.rb
@@ -1,9 +1,5 @@
 Rails.application.routes.draw do
+  devise_for :users, controllers: { registrations: 'registrations' }
 
-  devise_for :users, :controllers => { registrations: 'registrations'}
-  devise_scope :user do  
-root "devise/sessions#new"  
-end
-  #devise_for :users, :controllers => { registrations: 'registrations'}
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
+  root 'devise/sessions#new'
 end
