+++ config/webpack/environment.js
@@ -1,16 +1,11 @@
-const { environment } = require('@rails/webpacker')
-// const { default: Popper } = require('popper.js')
-const webpack = require('webpack')
-environment.plugins.append('Provide',
-    new webpack.ProvidePlugin({
-        $: 'jquery',
-        jQuery: 'jquery',
-        Popper: ['popper.js', 'default']
-    })
-)
-module.exports = environment
+    const { environment } = require('@rails/webpacker')
 
-
-
-
-
+    const webpack = require('webpack')
+    environment.plugins.append('Provide',
+        new webpack.ProvidePlugin({
+         $: 'jquery',
+         jQuery: 'jquery',
+         Popper: ['popper.js', 'default']
+        });
+    );
+    module.exports = environment
